Files: refactored file structure and ensured that all TSX files are in PascalCase
