+++ src/UI/Display/Display.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import VideoPlayer from '../Video/VideoPlayer';
+
+function Display() {
+  return (
+    <>
+      <VideoPlayer height={115} width={150} />
+    </>
+  );
+}
+
+export default Display;
