+++ src/World.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import Display from './UI/Display/Display';
+import Scene from './World/Scene/Scene';
+
+function World() {
+  return (
+    <>
+      <Display />
+      <Scene />
+    </>
+  );
+}
+
+export default World;
