+++ src/World/Player/Player.tsx
@@ -1,19 +1,19 @@
 import { SphereProps, useSphere } from '@react-three/cannon';
 import { useFrame, useThree } from '@react-three/fiber';
 import React, { useEffect, useRef } from 'react';
-import { playerMovementControls } from './playerMovementControls';
+import { playerMovementControls } from './PlayerMovementControls';
 import playerMovementEmitter from './playerMovementEmitter';
-import { PlayerPosition, PlayerVelocity } from '../allTypes';
+import { PlayerPosition, PlayerVelocity } from '../../allTypes';
 import { roundEntriesInVector, round } from './playerMovementUtils';
 import { Vector3 } from 'three';
-import { PlayerConstants, MathConstants } from '../constants';
+import { PlayerConstants, MathConstants } from '../../constants';
 
+/**
+ * Defines a custom player object (a default Cannon.js sphere and handles player movement)
+ * Movement works by examining keypresses and updating the player's velocity (handled by Cannon.js)
+ * according to keypress logic defined in ./playerMovementControls.ts.
+ */
 export default function Player(props: SphereProps) {
-  /**
-   * Defines a custom player object (a default Cannon.js sphere and handles player movement)
-   * Movement works by examining keypresses and updating the player's velocity (handled by Cannon.js)
-   * according to keypress logic defined in ./playerMovementControls.ts.
-   */
   const { forward, backward, left, right, jump } = playerMovementControls();
   const { camera } = useThree();
   const [playerRef, setPlayerRef] = useSphere(() => ({
@@ -30,7 +30,7 @@ export default function Player(props: SphereProps) {
 
   const currentVelocityVector = useRef<PlayerVelocity>({ x: 0, y: 0, z: 0 });
   useEffect(() => {
-    setPlayerRef.velocity.subscribe(playerVelocity => {
+    setPlayerRef.velocity.subscribe((playerVelocity) => {
       currentVelocityVector.current.x = round(
         playerVelocity[0],
         MathConstants.roundingPrecision
