+++ src/World/Scene/Floor.tsx
@@ -1,6 +1,6 @@
 import { usePlane } from '@react-three/cannon';
 import React from 'react';
-import { Colors } from '../constants';
+import { Colors } from '../../constants';
 
 export default function Floor() {
   /**
