+++ src/World/Scene/Skybox.tsx
@@ -1,6 +1,6 @@
 import { useThree } from '@react-three/fiber';
 import { CubeTextureLoader } from 'three';
-import { SkyboxImages } from '../constants';
+import { SkyboxImages } from '../../constants';
 
 /**
  * Applies a textured skybox.
